pyproject.toml
[project]
name = "aws-rekognition"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.40.55",
    "pillow>=12.0.0",
    "pydantic-settings>=2.11.0",
    "requests>=2.32.5",
]
